+++ src/utils.h
@@ -84,8 +84,8 @@ int CAN_SEE(struct char_data *s, struct char_data *o);
 #define GET_LEVEL(ch, i)   ((ch)->player.level[(i)])
 
 #define GET_CLASS_TITLE(ch, class, lev)   ((ch)->player.sex ?  \
-   (((ch)->player.sex == 1) ? titles[(class)][(lev)].title_m : \
-    titles[(class)][(lev)].title_f) : titles[(class)][(lev)].title_m)
+    (((ch)->player.sex == 1) ? titles[(class)][(int)(lev)].title_m : \
+    titles[(class)][(int)(lev)].title_f) : titles[(class)][(int)(lev)].title_m)
 
 #define GET_REQ(i) (i<2  ? "Awful" :(i<4  ? "Bad"     :(i<7  ? "Poor"      :\
 (i<10 ? "Average" :(i<14 ? "Fair"    :(i<20 ? "Good"    :(i<24 ? "Very good" :\
